requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.53.1
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pyparsing==3.1.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3